# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/seasons": {
            "get": {
                "tags": [
                    "Seasons"
                ],
                "summary": "Get All Seasons",
                "description": "Get Seasons details",
                "operationId": "GetSeasonsDetails",
                "responses": {
                    "200": {
                        "description": "Fetched successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Seasons"
                ],
                "summary": "Register a season",
                "description": "Register a season",
                "operationId": "registerSeasons",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "endDate"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "description": {
                                        "type": "text"
                                    },
                                    "startDate": {
                                        "type": "text"
                                    },
                                    "endDate": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully registered",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/seasons/{season}": {
            "get": {
                "tags": [
                    "Seasons"
                ],
                "summary": "Get A Single Season",
                "description": "Get Season details",
                "operationId": "GetSeasonDetails",
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fetched successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Seasons"
                ],
                "summary": "Update a season",
                "description": "update a season",
                "operationId": "UpdateSeason",
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "description": {
                                        "type": "text"
                                    },
                                    "startDate": {
                                        "type": "text"
                                    },
                                    "endDate": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Seasons"
                ],
                "summary": "Delete A Single Season",
                "description": "Delete Season details",
                "operationId": "DeleteSeasonDetails",
                "parameters": [
                    {
                        "name": "season",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}